Move the FPM runtime files after merging the separate repository
