--- src/FpmRuntime/.gitattributes
@@ -1,11 +1,0 @@
-# Do not export those files in the Composer archive (lighter dependency)
-.github/ export-ignore
-tests/ export-ignore
-.gitattributes export-ignore
-.gitignore export-ignore
-phpcs.xml export-ignore
-phpstan.neon export-ignore
-phpunit.xml export-ignore
-
-# Auto detect text files and perform LF normalization
-* text=auto
