--- src/FpmRuntime/.github/ISSUE_TEMPLATE/bug_report.md
@@ -1,16 +1,0 @@
----
-name: Bug report
-about: Report a bug in Bref
-title: ''
-labels: bug
-assignees: ''
-
----
-
-**Description:**
-
-<!-- A clear and concise description of what the bug is. -->
-
-**How to reproduce:**
-
-<!-- If you think that can help, please detail the version of Bref, which layer, and which AWS service you use. -->
