--- src/FpmRuntime/.github/ISSUE_TEMPLATE/config.yml
@@ -1,11 +1,0 @@
-blank_issues_enabled: true
-contact_links:
-    - name: Community support
-      url: https://github.com/brefphp/bref/discussions?discussions_q=category%3ASupport
-      about: Please ask and answer questions in GitHub Discussions.
-    - name: Professional support
-      url: https://bref.sh/#enterprise
-      about: Get in touch with Null, the company behind Bref.
-    - name: Sponsor a feature
-      url: https://bref.sh/#enterprise
-      about: Sponsor the development of a new feature you want to see in Bref.
