--- src/FpmRuntime/.github/ISSUE_TEMPLATE/feature-or-improvement.md
@@ -1,10 +1,0 @@
----
-name: Feature or improvement
-about: Suggest a feature or improvement
-title: ''
-labels: enhancement
-assignees: ''
-
----
-
-
