--- src/FpmRuntime/.github/pull_request_template.md
@@ -1,5 +1,0 @@
-<!--
-Please explain the motivation behind the changes.
-
-In other words, explain **WHY** instead of **WHAT**.
--->
