--- src/FpmRuntime/.gitignore
@@ -1,6 +1,0 @@
-/vendor
-/.idea
-/.phpcs-cache
-/composer.lock
-.DS_Store
-.phpunit.result.cache
