--- src/FpmRuntime/phpstan.neon
@@ -1,7 +1,0 @@
-parameters:
-  level: 5
-  paths:
-    - src
-    - tests
-  excludePaths:
-    - %rootDir%/../../../tests/Functional/fpm/*
