+++ tests/FpmRuntime/Functional/FpmRuntimeTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Bref\FpmRuntime\Test\Functional;
+namespace Bref\Test\FpmRuntime\Functional;
 
 use DMS\PHPUnitExtensions\ArraySubset\ArraySubsetAsserts;
 use GuzzleHttp\Client;
